Extend docker-build step to support more build parameters
